+++ Thrower.UnitTests/RaiseTests.cs
@@ -24,6 +24,7 @@
 using NUnit.Framework;
 using System;
 using System.Globalization;
+using System.IO;
 using System.Threading.Tasks;
 
 namespace PommaLabs.Thrower.UnitTests
@@ -409,6 +410,100 @@ namespace PommaLabs.Thrower.UnitTests
             {
                 Assert.AreEqual("Pino", ex.Message);
                 throw;
+            }
+        }
+
+        [Test]
+        public void ExtendedConstructor_FileNotFoundException()
+        {
+            const string msg = "Test MSG";
+            const string path = "Test PATH";
+
+            try
+            {
+                Raise<FileNotFoundException>.If(false, msg, path);
+            }
+            catch (Exception ex)
+            {
+                Assert.Fail(ex.Message);
+            }
+
+            try
+            {
+                Raise<FileNotFoundException>.If(true, msg, path);
+            }
+            catch (FileNotFoundException fex)
+            {
+                Assert.That(fex.Message, Is.EqualTo(msg));
+                Assert.That(fex.FileName, Is.EqualTo(path));
+            }
+            catch (Exception ex)
+            {
+                Assert.Fail(ex.Message);
+            }
+        }
+
+        [Test]
+        public void ExtendedConstructor_FileNotFoundException_Not()
+        {
+            const string msg = "Test MSG";
+            const string path = "Test PATH";
+
+            try
+            {
+                Raise<FileNotFoundException>.IfNot(true, msg, path);
+            }
+            catch (Exception ex)
+            {
+                Assert.Fail(ex.Message);
+            }
+
+            try
+            {
+                Raise<FileNotFoundException>.IfNot(false, msg, path);
+            }
+            catch (FileNotFoundException fex)
+            {
+                Assert.That(fex.Message, Is.EqualTo(msg));
+                Assert.That(fex.FileName, Is.EqualTo(path));
+            }
+            catch (Exception ex)
+            {
+                Assert.Fail(ex.Message);
+            }
+        }
+
+        [Test]
+        public void ExtendedConstructor_TooManyArgs()
+        {
+            try
+            {
+                Raise<ArgumentException>.If(true, 1, "snau", 3.14M);
+            }
+            catch (ArgumentException aex)
+            {
+                Assert.Pass(aex.Message);
+            }
+            catch (Exception ex)
+            {
+                Assert.Fail(ex.Message);
+            }
+        }
+
+        [Test]
+        public void ExtendedConstructor_TooManyArgs_Not()
+        {
+            try
+            {
+                Raise<ArgumentException>.IfNot(false, 1, "SNAFU", 3.14M);
+            }
+            catch (ArgumentException aex)
+            {
+                Assert.Pass(aex.Message);
+            }
+            catch (Exception ex)
+            {
+                Assert.Fail(ex.Message);
             }
         }